fixing route security

+++ app/controllers/lens_controller.rb
@@ -67,22 +67,26 @@ class LensController < ApplicationController
   patch "/lens/:id" do
     redirect_if_not_logged_in
     if len = Len.find_by(id: params[:id])
-      len.update(
-        make: params[:len][:make],
-        model: params[:len][:model],
-        focal_range: params[:len][:focal_range],
-        min_aperture: params[:len][:min_aperture],
-        max_aperture: params[:len][:max_aperture],
-        camera_id: params[:len][:camera_id],
-        user_id: params[:len][:user_id],
-        updated_at: Time.now
-      )
-      if len.save
-        redirect to "/lens/#{len.id}"
+      if len.user_id == session[:user_id]
+        len.update(
+          make: params[:len][:make],
+          model: params[:len][:model],
+          focal_range: params[:len][:focal_range],
+          min_aperture: params[:len][:min_aperture],
+          max_aperture: params[:len][:max_aperture],
+          camera_id: params[:len][:camera_id],
+          user_id: params[:len][:user_id],
+          updated_at: Time.now
+        )
+        if len.save
+          redirect to "/lens/#{len.id}"
+        else
+          @errors = len.errors.full_messages
+          @len = len
+          erb :"/lens/edit"
+        end
       else
-        @errors = len.errors.full_messages
-        @len = len
-        erb :"/lens/edit"
+        not_authorized
       end
     else
       not_authorized
@@ -92,9 +96,13 @@ class LensController < ApplicationController
   delete "/lens/:id" do
     redirect_if_not_logged_in
     if len = Len.find_by(id: params[:id])
-      flash[:message] = "#{len.make} - #{len.model} deleted successfully."
-      len.destroy
-      redirect to "/lens"
+      if len.user_id == session[:user_id]
+        flash[:message] = "#{len.make} - #{len.model} deleted successfully."
+        len.destroy
+        redirect to "/lens"
+      else
+        not_authorized
+      end
     else
       not_authorized
     end
