+++ app/controllers/users_controller.rb
@@ -58,15 +58,19 @@ class UsersController < ApplicationController
   patch "/users/:id" do
     redirect_if_not_logged_in
     if user = User.find_by(id: params[:id])
-      user.updated_at = Time.now
-      user.update(params[:user])
-      if user.save
-        session.clear
-        flash[:message] = "Account updated successfully."
-        redirect to "/login"
+      if user.id == session[:user_id]
+        user.updated_at = Time.now
+        user.update(params[:user])
+        if user.save
+          session.clear
+          flash[:message] = "Account updated successfully."
+          redirect to "/login"
+        else
+          @errors = user.errors.full_messages
+          erb :"/users/edit"
+        end
       else
-        @errors = user.errors.full_messages
-        erb :"/users/edit"
+        not_authorized
       end
     else
       not_authorized
@@ -76,41 +80,44 @@ class UsersController < ApplicationController
   delete "/users/:id" do
     redirect_if_not_logged_in
     if user = User.find_by(id: params[:id])
-
-      rolls = Roll.all
-      cameras = Camera.all
-      photos = Photo.all
-      lens = Len.all 
+      if user.id == session[:user_id]
+        rolls = Roll.all
+        cameras = Camera.all
+        photos = Photo.all
+        lens = Len.all 
+        
+        rolls.each do |roll|
+          if roll.user_id == user.id
+          roll.destroy
+          end
+        end
       
-      rolls.each do |roll|
-        if roll.user_id == user.id
-        roll.destroy
+        cameras.each do |camera|
+          if camera.user_id == user.id
+            camera.destroy
+          end
         end
-      end
-    
-      cameras.each do |camera|
-        if camera.user_id == user.id
-          camera.destroy
+      
+        photos.each do |photo|
+          if photo.user_id == user.id
+            photo.destroy
+          end
         end
-      end
-    
-      photos.each do |photo|
-        if photo.user_id == user.id
-          photo.destroy
+      
+        lens.each do |len|
+          if len.user_id == user.id
+            len.destroy
+          end
         end
-      end
-    
-      lens.each do |len|
-        if len.user_id == user.id
-          len.destroy
-        end
-      end
 
-      session.clear
-      user.destroy
-      flash[:message] = "Account deleted successfully."
+        session.clear
+        user.destroy
+        flash[:message] = "Account deleted successfully."
 
-      redirect to "/signup"
+        redirect to "/signup"
+      else
+        not_authorized
+      end
     else
       not_authorized
     end
